changes over weekend
Don't get mad at me by the little things that happened. I got horribly sick (and still horribly ill), I'll be most likely attending school tomorrow, just to work on the website as it's the final stretch very soon.

What I did manage to finish was with the sign up button on the Home page, now it will center now matter where you are (didn't need to use flex nor grid).

The mistake here is you told that h4 where is should be at all cost and assume the screen is always the same size.

Couldn't work on the video part. Will work on it later.

Worked on the firebase, constant error on "getProvider" but nearly got login system ready and once the error is fixed. I will be working on the sign up and then we will start on the chat system (hopefully coming by Friday)

I know this is a long documentation but its important to take notes the changes that have been made.