Final commit
We did it.

Months of work is finally over, we have finished the project with barely anytime to finish.

Though it's one huge mess of a website, it's our website and we learned something out of it. (I hope so).

Leó logging out of this system for the last time, as well with Ívan.

Co-Authored-By: GabrielMexico <118809024+GabrielMexico@users.noreply.github.com>